requirements.txt
python-telegram-bot==21.0.1
aiohttp==3.9.1
python-dotenv==1.0.0
apscheduler==3.10.4